requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
awscli==1.22.58
blinker==1.4
boto==2.49.0
boto3==1.21.3
botocore==1.24.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distlib==0.3.4
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
filelock==3.4.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==16.10.1
Jinja2==2.10
jmespath==0.10.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
olefile==0.45.1
PAM==0.4.2
Pillow==5.1.0
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.19.2
pyxdg==0.25
PyYAML==5.4.1
requests==2.18.4
requests-unixsocket==0.1.5
roman==2.0.0
rsa==4.7.2
s3transfer==0.5.1
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing-extensions==4.1.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
virtualenv==20.13.1
zipp==3.6.0
zope.interface==4.3.2
